start the focus moving and move the containers actions
