+++ src/enum_key.rs
@@ -1,10 +1,12 @@
 use crate::container::*;
+use crate::layout::*;
 use crate::windows::WindowsConf;
 use std::cmp::PartialEq;
 
 pub enum Action {
     AddPane,
     DeletePane,
+    MoveFocus,
 }
 
 pub struct KeyAction {
@@ -22,7 +24,10 @@ impl KeyAction {
                     .send(ChildToParent::AddChild(Container::MiniCont(minicont)))
                     .unwrap();
             }
-            Action::DeletePane => {}
+            Action::DeletePane => {},
+            Action::MoveFocus => {
+                config.get_sender().send(ChildToParent::MoveFocus(MoveDir::Left)).unwrap();
+            }
         }
     }
 }
